식단조회시 리뷰평점수정

+++ app-public-api/src/main/java/co/kurrant/app/public_api/service/impl/FoodServiceImpl.java
@@ -319,7 +319,7 @@ public class FoodServiceImpl implements FoodService {
 
         for (DailyFood dailyFood : dailyFoodList) {
 
-            Integer reviewAverage = qReviewRepository.findAllByFoodIdPageableLess(dailyFood.getFood().getId());
+            double reviewAverage = qReviewRepository.findAllByFoodIdPageableLess(dailyFood.getFood().getId());
 
             DiscountDto discountDto = OrderUtil.checkMembershipAndGetDiscountDto(user, spot.getGroup(), spot, dailyFood);
             DailyFoodDto dailyFoodDto = dailyFoodMapper.toDto(spot.getId(), dailyFood, discountDto, dailyFoodCountMap.get(dailyFood), userRecommendList, reviewAverage);
